# apps/docs/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MOCA API",
    "version": "1.0.0",
    "description": "The unified public API of the **Museum of Crypto Art** \u2014 collections, artworks (with original-ratio media), 3D exhibition rooms, and the **Art DeCC0s** knowledge base, aggregated behind a single API key.\n\nAll endpoints (except the index) require a MOCA API key sent as `X-API-Key: <key>` or `Authorization: Bearer <key>`.\n\nResponses use the Directus-style envelope: `{ \"data\": \u2026, \"meta\": \u2026 }` on success and `{ \"errors\": [{ \"message\", \"extensions\": { \"code\" } }] }` on failure.",
    "contact": {
      "name": "Museum of Crypto Art",
      "url": "https://museumofcryptoart.com"
    }
  },
  "servers": [
    {
      "url": "https://api.moca.qwellco.de",
      "description": "Production"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/v1/guide/exhibitions/{id}": {
      "get": {
        "tags": [
          "Guide"
        ],
        "operationId": "guideGetExhibition",
        "summary": "The exhibition context document",
        "security": [],
        "description": "The enriched context the guide reasons over: every room with its architect and description, every artwork with its artist \u2014 plus distinct artist/architect lists and Cortex question starters.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The exhibition id from the world builder."
          }
        ],
        "responses": {
          "200": {
            "description": "The context document",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/GuideContext"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Your MOCA API key."
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Your MOCA API key as a bearer token."
      }
    },
...
    "responses": {
...
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
...
    },
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                },
                "extensions": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "GuideContextArtwork": {
        "type": "object",
        "properties": {
          "id": {
            "type": [
              "integer",
              "null"
            ],
            "description": "Museum artwork id (nfts.id), when known."
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          },
          "artist": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ],
            "description": "Artwork description from museum data, clamped."
          },
          "collection": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "GuideContextRoom": {
        "type": "object",
        "properties": {
          "uid": {
            "type": "string",
            "description": "Placement uid from the exhibition export."
          },
          "id": {
            "type": [
              "integer",
              "null"
            ],
            "description": "Museum room id."
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          },
          "architect": {
            "type": [
              "string",
              "null"
            ],
            "description": "The room\u2019s architect, from museum data."
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "series": {
            "type": [
              "string",
              "null"
            ]
          },
          "artworks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GuideContextArtwork"
            }
          }
        }
      },
      "GuideContext": {
        "type": "object",
        "description": "The enriched context document the museum guide reasons over \u2014 the authoritative record of what hangs where in a spawned exhibition.",
        "properties": {
          "id": {
            "type": "string",
            "description": "Stable exhibition id from the world builder."
          },
          "name": {
            "type": "string"
          },
          "registeredAt": {
            "type": "string",
            "format": "date-time"
          },
          "generator": {
            "type": [
              "string",
              "null"
            ]
          },
          "rooms": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GuideContextRoom"
            }
          },
          "artists": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Distinct exhibited artist names."
          },
          "architects": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Distinct room architect names."
          },
          "counts": {
            "type": "object",
            "properties": {
              "rooms": {
                "type": "integer"
              },
              "artworks": {
                "type": "integer"
              },
              "artists": {
                "type": "integer"
              }
            }
          },
          "starters": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Cortex-suggested visitor questions (filled in asynchronously after registration)."
          }
        }
      },
...
    }
  }
}